==> 2_Loop_for_all.Rmd <==
---
title: "Untitled"
output: html_document
date: '2022-03-04'
editor_options: 
  chunk_output_type: console
---
Read in packages:
```{r}
library(tidyverse)
library(dplyr)
library(DBI)
library(odbc)
library(haven)
```


Read in data:
```{r}
#Etablering af forbindelse fra R til database.
con <- dbConnect(odbc(), "Forskerdatabase")

#Dataframe with blood tests and corresponding NPU code
npus <- colnames(read.csv("F:/Projekter/FSEID00001565/Data/Subprojects/Marie_IBD_clinical_biomarker/NPU-koder.txt"))

npu_df <- data.frame(name = c("F_cal", "CRP", "Hemoglobin", "Leukocytes", "Neutrophilocytes", "Lymphocytes", "Monocytes", "Basophiles", "Thrombocytes", "Iron", "VitB12", "Folate", "ALAT", "ASAT", "Bilirubin", "ESR2", "Eosinophiles1", "Albumin1", "Vitamin D1", "Neutrophilocytes2"), npu = c(npus[c(3,2,4,7:10, 13:14, 17:19)], "NPU19651", "NPU19654", "NPU01370", npus[c(6, 11, 15)], "NPU10267", "NPU02902"))

```

Summary statistics:
```{r}
diag1 <- read.csv("F:/Projekter/FSEID00001565/Data/Subprojects/Marie_IBD_clinical_biomarker/temp/IBD_diagnoses.txt", sep=" ")
cpr_reg <- read.csv("F:/Projekter/FSEID00001565/Data/Subprojects/Marie_IBD_clinical_biomarker/temp/Birth_and_gender.txt", sep= " ")
dar <- read.csv("F:/Projekter/FSEID00001565/Data/Subprojects/Marie_IBD_clinical_biomarker/temp/Death_times.txt", sep= " ")

#IBD cohort:
diag <- diag1 %>% filter(In_DK %in% c(1, NA))
diag <- left_join(diag, cpr_reg, by="lbnr")
diag <- diag %>% filter(greenland != 1)
diag <- left_join(diag, dar, by="lbnr")
diag <- diag %>% filter(as.Date(date_of_death)>as.Date("2007-12-31") | is.na(date_of_death))
table(diag$IBD)
diag <- diag %>% mutate(age =  as.numeric(as.Date(D_inddto) - as.Date(D_FODDATO))/365.25)

diag_CD <- diag %>% filter(IBD=="CD")
summary(diag_CD)
dim(diag_CD %>% filter(C_KON == "K"))[1]/dim(diag_CD)[1]

diag_UC <- diag %>% filter(IBD=="UC")
summary(diag_UC)
dim(diag_UC %>% filter(C_KON == "K"))[1]/dim(diag_UC)[1]
rm(diag_CD, diag_UC, diag1)

#Controls: Everyone in CPR register minus IBD, Greenland and persons dead before 2008 -> no measures in lab database!
controls <- cpr_reg %>% filter(greenland != 1) %>% filter(!(lbnr %in% diag$lbnr))
controls <- left_join(controls, dar, by="lbnr")
controls <- controls %>% filter(as.Date(date_of_death)>as.Date("2007-12-31") | is.na(date_of_death))
dim(controls)
dim(controls %>% filter(C_KON == "K"))[1]/dim(controls)[1]
```

#Matching Loop - both CD and UC:
```{r}
for (k in 1:dim(npu_df)[1]){
  lab_database <- dbGetQuery(con, paste0("SELECT * FROM FSEID00001565.LAB_LAB_DM_FORSKER WHERE ANALYSISCODE = '", npu_df[k, "npu"], "'"))
  
  # Add lbnr
  pseudo_cpr <- read_sas("F:/Projekter/FSEID00001565/InputData/fseid00001565_lbnr_key.sas7bdat")
  colnames(pseudo_cpr)[1] <- "PATIENT_CPR_ENCRYPTED"
  lab_database <- left_join(lab_database, pseudo_cpr, by="PATIENT_CPR_ENCRYPTED")
  lab_database <- lab_database %>% select(lbnr, LABORATORIUM_IDCODE, SAMPLINGDATE, UNIT, VALUE, REKVIRENT_IDTYPE)
  lab_database <- lab_database %>% filter(REKVIRENT_IDTYPE == "ydernummer")
  rm(pseudo_cpr)
  
  #Merge with IBD diagnoses
  diag1 <- read.csv("F:/Projekter/FSEID00001565/Data/Subprojects/Marie_IBD_clinical_biomarker/temp/IBD_diagnoses.txt", sep=" ")
  lab_database <- left_join(lab_database, diag1, by="lbnr")
  rm(diag1)

  #Keep only those, that were in DK 1 year before diagnosis
  lab_database <- lab_database %>% filter(In_DK %in% c(1, NA))

  #Merge with death register
  dar <- read.csv("F:/Projekter/FSEID00001565/Data/Subprojects/Marie_IBD_clinical_biomarker/temp/Death_times.txt", sep= " ")
  lab_database <- left_join(lab_database, dar, by="lbnr")
  rm(dar)

  #Add birthday and sex
  cpr_reg <- read.csv("F:/Projekter/FSEID00001565/Data/Subprojects/Marie_IBD_clinical_biomarker/temp/Birth_and_gender.txt", sep= " ")
  lab_database <- left_join(lab_database, cpr_reg, by="lbnr")
  rm(cpr_reg)

  #Remove those, that live in Greenland
  lab_database <- lab_database %>% filter(greenland != 1) %>% transform(SAMPLINGDATE = as.Date(SAMPLINGDATE), D_FODDATO = as.Date(D_FODDATO), D_inddto = as.Date(D_inddto), date_of_death = as.Date(date_of_death))

  # Keep only measures before diagnosis date
  lab_IBD <- lab_database %>% filter(!is.na(IBD) & SAMPLINGDATE<= D_inddto)
  lab_control <- lab_database %>% filter(is.na(IBD))

  #Gather all in data frame:
  lab_total <- rbind(lab_IBD, lab_control)
  rm(lab_database, lab_control)
  
  lab_UC <- lab_IBD %>% filter(IBD =="UC") 
  lab_CD <- lab_IBD %>% filter(IBD =="CD")
  rm(lab_IBD)

#Start with UC
  cases <- unique(lab_UC$lbnr)

# Find matched controls:
no_match <- c()
df_controls_UC <- data.frame()
for (i in cases){
  case_i <- lab_UC %>% filter(lbnr==i)
  ev_date <- case_i[1, "D_inddto"]
  controls <- lab_total

  # Remove samples from after diagnosis time, those with different sex and those with different age (+- 5 years)
  controls <-  controls %>% filter(SAMPLINGDATE <= ev_date & C_KON == case_i[1, "C_KON"] & D_FODDATO <= case_i[1, "D_FODDATO"] + 5*365.25 & D_FODDATO >= case_i[1, "D_FODDATO"] - 5*365.25)

  # Remove those who gets a diagnosis before or around (+- 0.5 year) diagnosis time and those who dies/emigrates/disappeared before or around (+- 0.5 year) diagnosis time
  remove <- controls %>% filter((!is.na(IBD)&D_inddto < ev_date - 0.5*365.25) | (!is.na(date_of_death)&date_of_death < ev_date - 0.5*365.25))
  controls <- controls %>% filter(!(lbnr %in% remove$lbnr))
  rm(remove)

  #Select 5 random controls for each sampling:
  df_controls <- data.frame()
  for (j in 1:dim(case_i)[1]){
    # Remove those with different lab and keep those with similar sampling date (+- 0.5 year)
    sub_controls <- controls %>% filter(LABORATORIUM_IDCODE %in% case_i[j, "LABORATORIUM_IDCODE"] & SAMPLINGDATE >= case_i[j, "SAMPLINGDATE"]- 0.5*365.25 & SAMPLINGDATE <= case_i[j, "SAMPLINGDATE"]+ 0.5*365.25)

    if(dim(sub_controls)[1]==0){
      no_match <- append(no_match, case_i[j, "lbnr"])
    }

    if(dim(sub_controls)[1]>0){
      # Select 5 random
    if(dim(sub_controls)[1]>5){
      sub_controls <- sample_n(sub_controls, 5)
    }

    # Change event and event date
    sub_controls$IBD <- "Control"
    sub_controls$D_inddto <- ev_date
    sub_controls$corresponding_case <- i

    #Couple to data frame
    df_controls <- rbind(df_controls, sub_controls)
    }
  }

  # Couple to data frame:
  df_controls_UC <- rbind(df_controls_UC, df_controls)
  rm(controls, df_controls, case_i, sub_controls)
}


# Couple controls to cases
lab_UC_match <- lab_UC %>% filter(!(lbnr %in% no_match))
lab_UC_match$corresponding_case <- NA
test_df_UC <- rbind(lab_UC_match, df_controls_UC) %>% mutate(sampling_to_diag = as.Date(D_inddto)-as.Date(SAMPLINGDATE))

# Save dataframe and rename:
write.table(test_df_UC, file=paste0("F:/Projekter/FSEID00001565/Data/Subprojects/Marie_IBD_clinical_biomarker/", npu_df[k, "name"],"_UC_onlyGP.txt"))

rm(lab_UC_match, lab_UC, test_df_UC, df_controls_UC, no_match, cases)


  ##############################################################################################
#Repeat for CD:

  cases <- unique(lab_CD$lbnr)

  # Find matched controls:
  no_match <- c()
  df_controls_CD <- data.frame()
  for (i in cases){
    case_i <- lab_CD %>% filter(lbnr==i)
    ev_date <- case_i[1, "D_inddto"]
    controls <- lab_total
    
    # Remove samples from after diagnosis time, those with different gender and those with different age (+- 5 years)
    controls <-  controls %>% filter(SAMPLINGDATE <= ev_date & C_KON == case_i[1, "C_KON"] & D_FODDATO <= case_i[1, "D_FODDATO"] + 5*365.25 & D_FODDATO >= case_i[1, "D_FODDATO"] - 5*365.25)

    # Remove those who gets a diagnosis before or around (+- 0.5 year) diagnosis time and those who dies/emigrates/disappeared before or around (+- 0.5 year) diagnosis time
    remove <- controls %>% filter((!is.na(IBD)&D_inddto < ev_date - 0.5*365.25) | (!is.na(date_of_death)&date_of_death < ev_date - 0.5*365.25))
    controls <- controls %>% filter(!(lbnr %in% remove$lbnr))
    rm(remove)
    
    #Select 5 random controls for each sampling:
    df_controls <- data.frame()
    for (j in 1:dim(case_i)[1]){
      # Remove those with different lab and keep those with similar sampling date (+- 0.5 year)
      sub_controls <- controls %>% filter(LABORATORIUM_IDCODE %in% case_i[j, "LABORATORIUM_IDCODE"] & SAMPLINGDATE >= case_i[j, "SAMPLINGDATE"]- 0.5*365.25 & SAMPLINGDATE <= case_i[j, "SAMPLINGDATE"]+ 0.5*365.25)
      
      if(dim(sub_controls)[1]==0){
        no_match <- append(no_match, case_i[j, "lbnr"])
      }
      
      if(dim(sub_controls)[1]>0){
        # Select 5 random
      if(dim(sub_controls)[1]>5){
        sub_controls <- sample_n(sub_controls, 5)
      }
      
      # Change event and event date
      sub_controls$IBD <- "Control"
      sub_controls$D_inddto <- ev_date
      sub_controls$corresponding_case <- i
      
      #Couple to data frame
      df_controls <- rbind(df_controls, sub_controls)
      }
    }

    # Couple to data frame:
    df_controls_CD <- rbind(df_controls_CD, df_controls)
    rm(controls, df_controls, case_i, sub_controls)
  }

  
  # Couple controls to cases
  lab_CD_match <- lab_CD %>% filter(!(lbnr %in% no_match))
  lab_CD_match$corresponding_case <- NA
  test_df_CD <- rbind(lab_CD_match, df_controls_CD) %>% mutate(sampling_to_diag = as.Date(D_inddto)-as.Date(SAMPLINGDATE))
  
  # Save dataframe and rename:
  write.table(test_df_CD, file=paste0("F:/Projekter/FSEID00001565/Data/Subprojects/Marie_IBD_clinical_biomarker/", npu_df[k, "name"],"_CD_onlyGP.txt"))
  
  rm(lab_CD_match, lab_CD, test_df_CD, df_controls_CD, no_match, cases, lab_total)
}

```

## Matching for sensitivity analysis: IBD based on 2 diagnoses
```{r}
for (k in 1:dim(npu_df)[1]){
  lab_database <- dbGetQuery(con, paste0("SELECT * FROM FSEID00001565.LAB_LAB_DM_FORSKER WHERE ANALYSISCODE = '", npu_df[k, "npu"], "'"))

  # Add lbnr
  pseudo_cpr <- read_sas("F:/Projekter/FSEID00001565/InputData/fseid00001565_lbnr_key.sas7bdat")
  colnames(pseudo_cpr)[1] <- "PATIENT_CPR_ENCRYPTED"
  lab_database <- left_join(lab_database, pseudo_cpr, by="PATIENT_CPR_ENCRYPTED")
  lab_database <- lab_database %>% select(lbnr, LABORATORIUM_IDCODE, SAMPLINGDATE, UNIT, VALUE, REKVIRENT_IDTYPE)
  lab_database <- lab_database %>% filter(REKVIRENT_IDTYPE == "ydernummer")
  rm(pseudo_cpr)
  
  #Merge with IBD diagnoses
  diag1 <- read.csv("F:/Projekter/FSEID00001565/Data/Subprojects/Marie_IBD_clinical_biomarker/temp/IBD_diagnoses.txt", sep=" ")
  lab_database <- left_join(lab_database, diag1, by="lbnr")
  rm(diag1)

  #Keep only those, that were in DK 1 year before diagnosis
  lab_database <- lab_database %>% filter(In_DK %in% c(1, NA))

  #Merge with death register
  dar <- read.csv("F:/Projekter/FSEID00001565/Data/Subprojects/Marie_IBD_clinical_biomarker/temp/Death_times.txt", sep= " ")
  lab_database <- left_join(lab_database, dar, by="lbnr")
  rm(dar)

  #Add birthday and sex
  cpr_reg <- read.csv("F:/Projekter/FSEID00001565/Data/Subprojects/Marie_IBD_clinical_biomarker/temp/Birth_and_gender.txt", sep= " ")
  lab_database <- left_join(lab_database, cpr_reg, by="lbnr")
  rm(cpr_reg)

  #Remove those, that live in Greenland
  lab_database <- lab_database %>% filter(greenland != 1) %>% transform(SAMPLINGDATE = as.Date(SAMPLINGDATE), D_FODDATO = as.Date(D_FODDATO), D_inddto = as.Date(D_inddto), date_of_death = as.Date(date_of_death), d_IBD_diag_reg2 = as.Date(d_IBD_diag_reg2))

  # Keep only measures before diagnosis date and remove non-specific (not two-diagnosis)
  lab_IBD <- lab_database %>% filter(!is.na(IBD_reg2) & SAMPLINGDATE<= D_inddto) %>% filter(IBD_specific == 1)
  lab_control <- lab_database %>% filter(is.na(IBD_reg2) | IBD_reg2=="")

  #Gather all in data frame:
  lab_total <- rbind(lab_IBD, lab_control)
  rm(lab_database, lab_control)
  
  lab_UC <- lab_IBD %>% filter(IBD_reg2 =="UC") 
  lab_CD <- lab_IBD %>% filter(IBD_reg2 =="CD")
  rm(lab_IBD)

  cases <- unique(lab_UC$lbnr)

  # Find matched controls:
  no_match <- c()
  df_controls_UC <- data.frame()
  for (i in cases){
    case_i <- lab_UC %>% filter(lbnr==i)
    ev_date <- case_i[1, "D_inddto"]
    match_date <- case_i[1, "d_IBD_diag_reg2"]
    controls <- lab_total

    # Remove samples from after diagnosis time, those with different gender and those with different age (+- 5 years)
    controls <-  controls %>% filter(SAMPLINGDATE <= ev_date & C_KON == case_i[1, "C_KON"] & D_FODDATO <= case_i[1, "D_FODDATO"] + 5*365.25 & D_FODDATO >= case_i[1, "D_FODDATO"] - 5*365.25)

    # Remove those who gets a diagnosis before or around (+- 0.5 year) matching time and those who dies/emigrates/disappeared before or around (+- 0.5 year) diagnosis time
    remove <- controls %>% filter((!is.na(IBD_reg2)&d_IBD_diag_reg2 < match_date - 0.5*365.25) | (!is.na(date_of_death)&date_of_death < match_date - 0.5*365.25))
    controls <- controls %>% filter(!(lbnr %in% remove$lbnr))
    rm(remove)

    #Select 5 random controls for each sampling:
    df_controls <- data.frame()
    for (j in 1:dim(case_i)[1]){
      # Remove those with different lab and keep those with similar sampling date (+- 0.5 year)
      sub_controls <- controls %>% filter(LABORATORIUM_IDCODE %in% case_i[j, "LABORATORIUM_IDCODE"] & SAMPLINGDATE >= case_i[j, "SAMPLINGDATE"]- 0.5*365.25 & SAMPLINGDATE <= case_i[j, "SAMPLINGDATE"]+ 0.5*365.25)

      if(dim(sub_controls)[1]==0){
        no_match <- append(no_match, case_i[j, "lbnr"])
      }

      if(dim(sub_controls)[1]>0){
        # Select 5 random
      if(dim(sub_controls)[1]>5){
        sub_controls <- sample_n(sub_controls, 5)
      }

      # Change event and event date
      sub_controls$IBD_reg2 <- "Control"
      sub_controls$D_inddto <- ev_date
      sub_controls$corresponding_case <- i

      #Couple to data frame
      df_controls <- rbind(df_controls, sub_controls)
      }
    }

    # Couple to data frame:
    df_controls_UC <- rbind(df_controls_UC, df_controls)
    rm(controls, df_controls, case_i, sub_controls)
  }


  # Couple controls to cases
  lab_UC_match <- lab_UC %>% filter(!(lbnr %in% no_match))
  lab_UC_match$corresponding_case <- NA
  test_df_UC <- rbind(lab_UC_match, df_controls_UC) %>% mutate(sampling_to_diag = as.Date(D_inddto)-as.Date(SAMPLINGDATE))

  # Save data frame and rename:
  write.table(test_df_UC, file=paste0("F:/Projekter/FSEID00001565/Data/Subprojects/Marie_IBD_clinical_biomarker/Sensitivity_" , npu_df[k, "name"],"_UC_onlyGP.txt"))

  rm(lab_UC_match, lab_UC, test_df_UC, df_controls_UC, no_match, cases)


  ##############################################################################################

  cases <- unique(lab_CD$lbnr)

  # Find matched controls:
  no_match <- c()
  df_controls_CD <- data.frame()
  for (i in cases){
    case_i <- lab_CD %>% filter(lbnr==i)
    ev_date <- case_i[1, "D_inddto"]
    match_date <- case_i[1, "d_IBD_diag_reg2"]
    controls <- lab_total
    
    # Remove samples from after diagnosis time, those with different gender and those with different age (+- 5 years)
    controls <-  controls %>% filter(SAMPLINGDATE <= ev_date & C_KON == case_i[1, "C_KON"] & D_FODDATO <= case_i[1, "D_FODDATO"] + 5*365.25 & D_FODDATO >= case_i[1, "D_FODDATO"] - 5*365.25)

    # Remove those who gets a diagnosis before or around (+- 0.5 year) matching time and those who dies/emigrates/disappeared before or around (+- 0.5 year) matching time
    remove <- controls %>% filter((!is.na(IBD_reg2)&d_IBD_diag_reg2 < match_date - 0.5*365.25) | (!is.na(date_of_death)&date_of_death < match_date - 0.5*365.25))
    controls <- controls %>% filter(!(lbnr %in% remove$lbnr))
    rm(remove)
    
    #Select 5 random controls for each sampling:
    df_controls <- data.frame()
    for (j in 1:dim(case_i)[1]){
      # Remove those with different lab and keep those with similar sampling date (+- 0.5 year)
      sub_controls <- controls %>% filter(LABORATORIUM_IDCODE %in% case_i[j, "LABORATORIUM_IDCODE"] & SAMPLINGDATE >= case_i[j, "SAMPLINGDATE"]- 0.5*365.25 & SAMPLINGDATE <= case_i[j, "SAMPLINGDATE"]+ 0.5*365.25)
      
      if(dim(sub_controls)[1]==0){
        no_match <- append(no_match, case_i[j, "lbnr"])
      }
      
      if(dim(sub_controls)[1]>0){
        # Select 5 random
      if(dim(sub_controls)[1]>5){
        sub_controls <- sample_n(sub_controls, 5)
      }
      
      # Change event and event date
      sub_controls$IBD_reg2 <- "Control"
      sub_controls$D_inddto <- ev_date
      sub_controls$corresponding_case <- i
      
      #Couple to data frame
      df_controls <- rbind(df_controls, sub_controls)
      }
    }

    # Couple to data frame:
    df_controls_CD <- rbind(df_controls_CD, df_controls)
    rm(controls, df_controls, case_i, sub_controls)
  }

  
  # Couple controls to cases
  lab_CD_match <- lab_CD %>% filter(!(lbnr %in% no_match))
  lab_CD_match$corresponding_case <- NA
  test_df_CD <- rbind(lab_CD_match, df_controls_CD) %>% mutate(sampling_to_diag = as.Date(D_inddto)-as.Date(SAMPLINGDATE))
  
  # Save dataframe and rename:
  write.table(test_df_CD, file=paste0("F:/Projekter/FSEID00001565/Data/Subprojects/Marie_IBD_clinical_biomarker/Sensitivity_", npu_df[k, "name"],"_CD_onlyGP.txt"))
  
  rm(lab_CD_match, lab_CD, test_df_CD, df_controls_CD, no_match, cases, lab_total)
}

```



# Charlson's co-morbidities (sensitivity analysis):
For each NPU code >
For each subject >
Compute Charlson index, going 9 years back in time. (and save all diagnoses to get an overview later)
```{r}

# Read in LPR:
lpr <- read_sas("F:/Projekter/FSEID00001565/Data/Prepared_basefiles/ibd_patients_lpr_diags.sas7bdat")

# Keep only diagtype A, B and G: 
lpr <- lpr %>% filter(C_diagtype %in% c("A", "B", "G") & (C_pattype == "0" | C_pattype == "1" | (C_pattype == "2" & D_inddto<= as.Date("2013/12/31", format="%Y/%m/%d")) | (C_pattype == "2" & C_indm != "1" & D_inddto > as.Date("2013/12/31", format="%Y/%m/%d"))))

for (k in 1:dim(npu_df)[1]){
  #Read in CD:
  df_CD <- read.csv(paste0("F:/Projekter/FSEID00001565/Data/Subprojects/Marie_IBD_clinical_biomarker/", npu_df[k, "name"], "_CD_onlyGP.txt"), sep=" ")
  df_CD <- df_CD %>% mutate(D_inddto=as.Date(D_inddto))
  df_CD <- df_CD %>% dplyr:: select(lbnr, LABORATORIUM_IDCODE, SAMPLINGDATE, VALUE, D_inddto, IBD, D_FODDATO, C_KON, sampling_to_diag)

  #Make new empty columns:
  df_CD$chgrp_1 <- NA
  df_CD$chgrp_2 <- NA
  df_CD$chgrp_3 <- NA
  df_CD$chgrp_4 <- NA
  df_CD$chgrp_5 <- NA
  df_CD$chgrp_6 <- NA
  df_CD$chgrp_7 <- NA
  df_CD$chgrp_8 <- NA
  df_CD$chgrp_9 <- NA
  df_CD$chgrp_10 <- NA
  df_CD$chgrp_11 <- NA
  df_CD$chgrp_12 <- NA
  df_CD$chgrp_13 <- NA
  df_CD$chgrp_14 <- NA
  df_CD$chgrp_15 <- NA
  df_CD$chgrp_16 <- NA
  df_CD$chgrp_17 <- NA
  df_CD$comorb_score <- NA

  #Loop for each subject:
  for (i in 1:dim(df_CD)[1]){
    nr <- df_CD[i, "lbnr"]
    tstop <- df_CD[i, "D_inddto"]

    lpr_sub <- lpr %>% filter(lbnr == nr & D_inddto <= tstop & D_inddto >= tstop-365.25*9)

    if (dim(lpr_sub)[1]>0){
      #Find diagnoses:
    for (j in lpr_sub$C_diag){
      ch_grp_1 <- 0
      ch_grp_2 <- 0
      ch_grp_3 <- 0
      ch_grp_4 <- 0
      ch_grp_5 <- 0
      ch_grp_6 <- 0
      ch_grp_7 <- 0
      ch_grp_8 <- 0
      ch_grp_9 <- 0
      ch_grp_10 <- 0
      ch_grp_11 <- 0
      ch_grp_12 <- 0
      ch_grp_13 <- 0
      ch_grp_14 <- 0
      ch_grp_15 <- 0
      ch_grp_16 <- 0
      ch_grp_17 <- 0

    ## CH diagnoses:
      if ((substr(j, 1, 4) %in% c('DI21','DI22','DI23') | substr(j, 1, 3) == '410')){
        ch_grp_1 <- 1
      } else if ((substr(j, 1, 4) == 'DI50' | substr(j, 1, 5) %in% c('DI110','DI130','DI132','42709','42710','42711','42719','42899','78249'))){
        ch_grp_2 <- 1
      } else if ((substr(j, 1, 4) %in% c('DI70','DI71','DI72','DI73','DI74','DI77') | substr(j, 1, 3) %in% c('440','441','442','443','444','445'))){
        ch_grp_3 <- 1
      } else if ((substr(j, 1, 4) %in% c('DI60','DI61','DI62','DI63','DI64','DI65','DI66','DI67','DI68','DI69','DG45','DG46') | substr(j, 1, 3) %in% c('430','431','432','433','434','435','436','437','438'))){
        ch_grp_4 <- 1
      } else if ((substr(j, 1, 4) %in% c('DF00','DF01','DF02','DF03','DG30') | substr(j, 1, 5) %in% c('DF051','29009','29010','29011','29012','29013','29014','29015','29016','29017','29018','29019','29309'))){
        ch_grp_5 <- 1
      } else if ((substr(j, 1, 4) %in% c('DJ40','DJ41','DJ42','DJ43','DJ44','DJ45','DJ46','DJ47','DJ60','DJ61','DJ62','DJ63','DJ64','DJ65','DJ66','DJ67') | substr(j, 1, 3) %in% c('490','491','492','493','515','516','517','518') | substr(j, 1, 5) %in% c('DJ684','DJ701','DJ703','DJ841','DJ920','DJ961','DJ982','DJ983'))){
        ch_grp_6 <- 1
      } else if ((substr(j, 1, 4) %in% c('DM05','DM06','DM08','DM09','DM30','DM31','DM32','DM33','DM34','DM35','DM36','DD86') | substr(j, 1, 3) %in% c('712','716','734','446') | substr(j, 1, 5) == '13599')){
        ch_grp_7 <- 1
      } else if ((substr(j, 1, 4) %in% c('DK25','DK26','DK27','DK28') | substr(j, 1, 3) %in% c('531','532','533','534') | substr(j, 1, 5) %in% c('DK221','53091','53098'))){
        ch_grp_8 <- 1
      } else if ((substr(j, 1, 4) %in% c('DB18','DK71','DK73','DK74') | substr(j, 1, 3) == '571' | substr(j, 1, 5) %in% c('DK700','DK701','DK702','DK703','DK709','DK760','57301','57304'))){
        ch_grp_9 <- 1
      } else if ((substr(j, 1, 5) %in% c('DE100','DE101','DE109','DE110','DE111','DE119','24900','24906','24907','24909','25000','25006','25007','25009'))){
        ch_grp_10 <- 1
      } else if ((substr(j, 1, 5) %in% c('DE102','DE103','DE104','DE105','DE106','DE107','DE108','DE112','DE113','DE114','DE115','DE116','DE117','DE118','24901','24902','24903','24904','24905','24908','25001','25002','25003','25004','25005','25008'))){
        ch_grp_11 <- 1
      } else if ((substr(j, 1, 4) %in% c('DG81','DG82') | substr(j, 1, 3) == '344')){
        ch_grp_12 <- 1
      } else if ((substr(j, 1, 4) %in% c('DI12','DI13','DN00','DN01','DN02','DN03','DN04','DN05','DN07','DN11','DN14','DN17','DN18','DN19','DQ61') | substr(j, 1, 3) %in% c('403','404','580','581','582','583','584','792') | substr(j, 1, 5) %in% c('59009','59319','75310','75311','75312','75313','75314','75315','75316','75317','75318','75319'))){
        ch_grp_13 <- 1
      } else if ((substr(j, 1, 3) %in% c('DC0','DC1','DC2','DC3','DC4','DC5','DC6','190','191','192','193','194','200','201','202','203','204','205','206','207') | substr(j, 1, 4) %in% c('DC70','DC71','DC72','DC73','DC74','DC75','DC81','DC82','DC83','DC84','DC85','DC88','DC90','DC91','DC92','DC93','DC94','DC95','DC96') | substr(j, 1, 2) %in% c('14','15','16','17','18') | substr(j, 1, 5) == '27559')){
        ch_grp_14 <- 1
      } else if ((substr(j, 1, 4) %in% c('DK72','DI85','4560') | substr(j, 1, 5) %in% c('DB150','DB160','DB162','DB190','DK704','DK766','07000','07002','07004','07006','07008','57300'))){
        ch_grp_15 <- 1
      } else if ((substr(j, 1, 4) %in% c('DC76','DC77','DC78','DC79','DC80') | substr(j, 1, 3) %in% c('195','196','197','198','199'))){
        ch_grp_16 <- 1
      } else if ((substr(j, 1, 4) %in% c('DB21','DB22','DB23','DB24') | substr(j, 1, 5) == '07983')){
        ch_grp_17 <- 1
      }
    }
    chi_score <- sum(sum(ch_grp_1, ch_grp_2, ch_grp_3, ch_grp_4, ch_grp_5, ch_grp_6, ch_grp_7, ch_grp_8), max(ch_grp_9*1, ch_grp_15*3), max(ch_grp_10*1, ch_grp_11*2), 2*sum(ch_grp_12, ch_grp_13), max(ch_grp_14*2, ch_grp_16*6), ch_grp_17*6)

    #Add the data frame:
    df_CD[i, 10:27] <- c(ch_grp_1, ch_grp_2, ch_grp_3, ch_grp_4, ch_grp_5, ch_grp_6, ch_grp_7, ch_grp_8, ch_grp_9, ch_grp_10, ch_grp_11, ch_grp_12, ch_grp_13, ch_grp_14, ch_grp_15, ch_grp_16, ch_grp_17, chi_score)

    rm(lpr_sub, ch_grp_1, ch_grp_2, ch_grp_3, ch_grp_4, ch_grp_5, ch_grp_6, ch_grp_7, ch_grp_8, ch_grp_9, ch_grp_10, ch_grp_11, ch_grp_12, ch_grp_13, ch_grp_14, ch_grp_15, ch_grp_16, ch_grp_17, chi_score, nr, tstop)

    } else {
      df_CD[i, 10:27] <- rep(0, 18)
    }
  rm(lpr_sub)
  }

  #Save data frame:
  write.table(df_CD, file=paste0("F:/Projekter/FSEID00001565/Data/Subprojects/Marie_IBD_clinical_biomarker/", npu_df[k, "name"],"_CD_comorb_onlyGP.txt"))
  rm(df_CD)


  ####################################################################################################################
  #UC:
  
    #Read in UC: 
  df_UC <- read.csv(paste0("F:/Projekter/FSEID00001565/Data/Subprojects/Marie_IBD_clinical_biomarker/", npu_df[k, "name"], "_UC_onlyGP.txt"), sep=" ")
  df_UC <- df_UC %>% mutate(D_inddto=as.Date(D_inddto))
  df_UC <- df_UC %>% dplyr:: select(lbnr, LABORATORIUM_IDCODE, SAMPLINGDATE, VALUE, D_inddto, IBD, D_FODDATO, C_KON, sampling_to_diag)

  #Make new empty columns:
  df_UC$chgrp_1 <- NA
  df_UC$chgrp_2 <- NA
  df_UC$chgrp_3 <- NA
  df_UC$chgrp_4 <- NA
  df_UC$chgrp_5 <- NA
  df_UC$chgrp_6 <- NA
  df_UC$chgrp_7 <- NA
  df_UC$chgrp_8 <- NA
  df_UC$chgrp_9 <- NA
  df_UC$chgrp_10 <- NA
  df_UC$chgrp_11 <- NA
  df_UC$chgrp_12 <- NA
  df_UC$chgrp_13 <- NA
  df_UC$chgrp_14 <- NA
  df_UC$chgrp_15 <- NA
  df_UC$chgrp_16 <- NA
  df_UC$chgrp_17 <- NA
  df_UC$comorb_score <- NA
  
  
  #Loop for each subject:
  for (i in 1:dim(df_UC)[1]){
    nr <- df_UC[i, "lbnr"]
    tstop <- df_UC[i, "D_inddto"]
    lpr_sub <- lpr %>% filter(lbnr == nr & D_inddto <= tstop & D_inddto >= tstop-365.25*9)
    
     if (dim(lpr_sub)[1]>0){
      #Find diagnoses:
    for (j in lpr_sub$C_diag){
      ch_grp_1 <- 0
      ch_grp_2 <- 0
      ch_grp_3 <- 0
      ch_grp_4 <- 0
      ch_grp_5 <- 0
      ch_grp_6 <- 0
      ch_grp_7 <- 0
      ch_grp_8 <- 0
      ch_grp_9 <- 0
      ch_grp_10 <- 0
      ch_grp_11 <- 0
      ch_grp_12 <- 0
      ch_grp_13 <- 0
      ch_grp_14 <- 0
      ch_grp_15 <- 0
      ch_grp_16 <- 0
      ch_grp_17 <- 0
      
    ## CH diagnoses:
      if ((substr(j, 1, 4) %in% c('DI21','DI22','DI23') | substr(j, 1, 3) == '410')){
        ch_grp_1 <- 1
      } else if ((substr(j, 1, 4) == 'DI50' | substr(j, 1, 5) %in% c('DI110','DI130','DI132','42709','42710','42711','42719','42899','78249'))){
        ch_grp_2 <- 1
      } else if ((substr(j, 1, 4) %in% c('DI70','DI71','DI72','DI73','DI74','DI77') | substr(j, 1, 3) %in% c('440','441','442','443','444','445'))){
        ch_grp_3 <- 1
      } else if ((substr(j, 1, 4) %in% c('DI60','DI61','DI62','DI63','DI64','DI65','DI66','DI67','DI68','DI69','DG45','DG46') | substr(j, 1, 3) %in% c('430','431','432','433','434','435','436','437','438'))){
        ch_grp_4 <- 1
      } else if ((substr(j, 1, 4) %in% c('DF00','DF01','DF02','DF03','DG30') | substr(j, 1, 5) %in% c('DF051','29009','29010','29011','29012','29013','29014','29015','29016','29017','29018','29019','29309'))){
        ch_grp_5 <- 1
      } else if ((substr(j, 1, 4) %in% c('DJ40','DJ41','DJ42','DJ43','DJ44','DJ45','DJ46','DJ47','DJ60','DJ61','DJ62','DJ63','DJ64','DJ65','DJ66','DJ67') | substr(j, 1, 3) %in% c('490','491','492','493','515','516','517','518') | substr(j, 1, 5) %in% c('DJ684','DJ701','DJ703','DJ841','DJ920','DJ961','DJ982','DJ983'))){
        ch_grp_6 <- 1
      } else if ((substr(j, 1, 4) %in% c('DM05','DM06','DM08','DM09','DM30','DM31','DM32','DM33','DM34','DM35','DM36','DD86') | substr(j, 1, 3) %in% c('712','716','734','446') | substr(j, 1, 5) == '13599')){
        ch_grp_7 <- 1
      } else if ((substr(j, 1, 4) %in% c('DK25','DK26','DK27','DK28') | substr(j, 1, 3) %in% c('531','532','533','534') | substr(j, 1, 5) %in% c('DK221','53091','53098'))){
        ch_grp_8 <- 1
      } else if ((substr(j, 1, 4) %in% c('DB18','DK71','DK73','DK74') | substr(j, 1, 3) == '571' | substr(j, 1, 5) %in% c('DK700','DK701','DK702','DK703','DK709','DK760','57301','57304'))){
        ch_grp_9 <- 1
      } else if ((substr(j, 1, 5) %in% c('DE100','DE101','DE109','DE110','DE111','DE119','24900','24906','24907','24909','25000','25006','25007','25009'))){
        ch_grp_10 <- 1
      } else if ((substr(j, 1, 5) %in% c('DE102','DE103','DE104','DE105','DE106','DE107','DE108','DE112','DE113','DE114','DE115','DE116','DE117','DE118','24901','24902','24903','24904','24905','24908','25001','25002','25003','25004','25005','25008'))){
        ch_grp_11 <- 1
      } else if ((substr(j, 1, 4) %in% c('DG81','DG82') | substr(j, 1, 3) == '344')){
        ch_grp_12 <- 1
      } else if ((substr(j, 1, 4) %in% c('DI12','DI13','DN00','DN01','DN02','DN03','DN04','DN05','DN07','DN11','DN14','DN17','DN18','DN19','DQ61') | substr(j, 1, 3) %in% c('403','404','580','581','582','583','584','792') | substr(j, 1, 5) %in% c('59009','59319','75310','75311','75312','75313','75314','75315','75316','75317','75318','75319'))){
        ch_grp_13 <- 1
      } else if ((substr(j, 1, 3) %in% c('DC0','DC1','DC2','DC3','DC4','DC5','DC6','190','191','192','193','194','200','201','202','203','204','205','206','207') | substr(j, 1, 4) %in% c('DC70','DC71','DC72','DC73','DC74','DC75','DC81','DC82','DC83','DC84','DC85','DC88','DC90','DC91','DC92','DC93','DC94','DC95','DC96') | substr(j, 1, 2) %in% c('14','15','16','17','18') | substr(j, 1, 5) == '27559')){
        ch_grp_14 <- 1
      } else if ((substr(j, 1, 4) %in% c('DK72','DI85','4560') | substr(j, 1, 5) %in% c('DB150','DB160','DB162','DB190','DK704','DK766','07000','07002','07004','07006','07008','57300'))){
        ch_grp_15 <- 1
      } else if ((substr(j, 1, 4) %in% c('DC76','DC77','DC78','DC79','DC80') | substr(j, 1, 3) %in% c('195','196','197','198','199'))){
        ch_grp_16 <- 1
      } else if ((substr(j, 1, 4) %in% c('DB21','DB22','DB23','DB24') | substr(j, 1, 5) == '07983')){
        ch_grp_17 <- 1
      }
    }
    chi_score <- sum(sum(ch_grp_1, ch_grp_2, ch_grp_3, ch_grp_4, ch_grp_5, ch_grp_6, ch_grp_7, ch_grp_8), max(ch_grp_9*1, ch_grp_15*3), max(ch_grp_10*1, ch_grp_11*2), 2*sum(ch_grp_12, ch_grp_13), max(ch_grp_14*2, ch_grp_16*6), ch_grp_17*6)
    
    #Add the data frame:
    df_UC[i, 10:27] <- c(ch_grp_1, ch_grp_2, ch_grp_3, ch_grp_4, ch_grp_5, ch_grp_6, ch_grp_7, ch_grp_8, ch_grp_9, ch_grp_10, ch_grp_11, ch_grp_12, ch_grp_13, ch_grp_14, ch_grp_15, ch_grp_16, ch_grp_17, chi_score)
    
    rm(lpr_sub, ch_grp_1, ch_grp_2, ch_grp_3, ch_grp_4, ch_grp_5, ch_grp_6, ch_grp_7, ch_grp_8, ch_grp_9, ch_grp_10, ch_grp_11, ch_grp_12, ch_grp_13, ch_grp_14, ch_grp_15, ch_grp_16, ch_grp_17, chi_score, nr, tstop)
    
    } else {
      df_UC[i, 10:27] <- rep(0, 18)
    }
    
  }
  #Save data frame: 
  write.table(df_UC, file=paste0("F:/Projekter/FSEID00001565/Data/Subprojects/Marie_IBD_clinical_biomarker/", npu_df[k, "name"],"_UC_comorb_onlyGP.txt"))
  rm(df_UC)
}

```

grp_1='Myocardial infarction'
grp_2='Congestive heart failure'
grp_3='Periphral vascular disease'
grp_4='Cerebrovascular disease'
grp_5='Dementia' 
grp_6='Chronic pulmonary disease' 
grp_7='Connective tissue diesease - Rheumatic disease' 
grp_8='Peptic ulcer disease' 
grp_9='Mild liver disease' 
grp_10='Diabetes without complications' 
grp_11='Diabetes with complications' 
grp_12='Paraplegia and hemiplegia'
grp_13='Renal disease' 
grp_14='Cancer' 
grp_15='Moderate or severe liver disease' 
grp_16='Metastatic carcinoma' 
grp_17='AIDS/HIV'
