==> 3_VitD_analysis.Rmd <==
---
title: "Untitled"
output: html_document
editor_options: 
  chunk_output_type: console
---

Read in packages:
```{r}
library(ggplot2)
library(tidyverse)
library(dplyr)
library(DBI)
library(odbc)
library(haven)
library(lme4)
library(lmerTest)
library(nlme)
library(ggpubr)
```


Load data:
```{r}
test_df_CD <- read.csv("F:/Projekter/FSEID00001565/Data/Subprojects/Marie_IBD_clinical_biomarker/Vitamin D1_CD_onlyGP.txt", sep=" ")
test_df_UC <- read.csv("F:/Projekter/FSEID00001565/Data/Subprojects/Marie_IBD_clinical_biomarker/Vitamin D1_UC_onlyGP.txt", sep=" ")

test_df_CD <- test_df_CD %>% mutate(IBD = fct_relevel(IBD, "Control", "CD"))

```

Check normality of the features and remove outliers:
```{r}
test_df_CD$measure_transformed <- as.numeric(test_df_CD$VALUE)
test_df_CD$VALUE[which(is.na(test_df_CD$measure_transformed))]

test_df_UC$measure_transformed <- as.numeric(test_df_UC$VALUE)
test_df_UC$VALUE[which(is.na(test_df_UC$measure_transformed))]

test_df_CD$VALUE <- gsub(",", ".", test_df_CD$VALUE)
test_df_UC$VALUE <- gsub(",", ".", test_df_UC$VALUE)

sub_CD <- test_df_CD %>% filter(as.numeric(VALUE)<10)
m_CD <- median(as.numeric(sub_CD$VALUE))
test_df_CD$VALUE <- gsub("<10", m_CD, test_df_CD$VALUE)
sub_CD <- test_df_CD %>% filter(as.numeric(VALUE)<12)
m_CD <- median(as.numeric(sub_CD$VALUE))
test_df_CD$VALUE <- gsub("<12", m_CD, test_df_CD$VALUE)
sub_CD <- test_df_CD %>% filter(as.numeric(VALUE)>375)
m_CD <- median(as.numeric(sub_CD$VALUE))
test_df_CD$VALUE <- gsub(">375", m_CD, test_df_CD$VALUE)
rm(sub_CD)

sub_UC <- test_df_UC %>% filter(as.numeric(VALUE)<10)
m_UC <- median(as.numeric(sub_UC$VALUE))
test_df_UC$VALUE <- gsub("<10", m_UC, test_df_UC$VALUE)
sub_UC <- test_df_UC %>% filter(as.numeric(VALUE)<12)
m_UC <- median(as.numeric(sub_UC$VALUE))
test_df_UC$VALUE <- gsub("<12", m_UC, test_df_UC$VALUE)
sub_UC <- test_df_UC %>% filter(as.numeric(VALUE)>375)
m_UC <- median(as.numeric(sub_UC$VALUE))
test_df_UC$VALUE <- gsub(">375", m_CD, test_df_UC$VALUE)
rm(sub_UC)

quantile(as.numeric(c(test_df_CD$VALUE, test_df_UC$VALUE)), probs=c(0.5, 0.25, 0.75))

#Look at distribution:
 p1 <-  ggplot(test_df_CD)+
    geom_histogram(aes(x=as.numeric(VALUE)), bins=30) + theme_classic()+ylab("Count") + xlab("Vitamin D2+D3")
 p2 <-  ggplot(test_df_CD)+
    geom_histogram(aes(x=log(as.numeric(VALUE))+1), bins=30)+ theme_classic()+ylab("Count") + xlab("Vitamin D2+D3 - log transformed")
  
 p3 <-  ggplot(test_df_UC)+
    geom_histogram(aes(x=as.numeric(VALUE)), bins=30)+ theme_classic()+ylab("Count") + xlab("Vitamin D2+D3")
 p4 <-  ggplot(test_df_UC)+
    geom_histogram(aes(x=log(as.numeric(VALUE))+1), bins=30)+ theme_classic()+ylab("Count") + xlab("Vitamin D2+D3 - log transformed")
 
 ggarrange(p1, p2)
 ggarrange(p3, p4)
 
 ggplot()+
    geom_histogram(aes(x=log(as.numeric(c(test_df_CD$VALUE, test_df_UC$VALUE)))+1), bins=30)+ theme_classic()+ylab("Count") + xlab("Vitamin D2+D3 - log transformed")
  
#Transform when necessary -> make a new column (measure_transformed)
test_df_CD$measure_transformed <- log(as.numeric(test_df_CD$VALUE)+1)
test_df_UC$measure_transformed <- log(as.numeric(test_df_UC$VALUE)+1)

#Remove outliers
  m_UC <- mean(test_df_UC$measure_transformed, na.rm = T)
  sd_UC <- sd(test_df_UC$measure_transformed, na.rm=T)
  outliers_UC <- 0
  
  m_CD <- mean(test_df_CD$measure_transformed, na.rm = T)
  sd_CD <- sd(test_df_CD$measure_transformed, na.rm = T)
  outliers_CD <- 0

  for (i in dim(test_df_UC)[1]){
   val <- test_df_UC[i, "measure_transformed"]
   
   if(val > m_UC+3*sd_UC | val < m_UC-3*sd_UC){
      test_df_UC[i, "measure_transformed"] <- NA
      outliers_UC <- outliers_UC +1
   }
  }
  
  for (i in dim(test_df_CD)[1]){
   val <- test_df_CD[i, "measure_transformed"]
   
   if(val > m_CD+3*sd_CD | val < m_CD-3*sd_CD){
      test_df_CD[i, "measure_transformed"] <- NA
      outliers_CD <- outliers_CD +1
   }
    }

```


For each feature, run a linear mixed effects model:
```{r}
test_df_CD <- test_df_CD %>% mutate(age =  as.numeric(as.Date(SAMPLINGDATE) - as.Date(D_FODDATO))/365.25)
test_df_UC <- test_df_UC %>% mutate(age =  as.numeric(as.Date(SAMPLINGDATE) - as.Date(D_FODDATO))/365.25)

#Overview of data
summary(test_df_UC %>% filter(IBD == "UC"))
dim(test_df_UC %>% filter(IBD == "UC", C_KON == "K"))[1]/dim(test_df_UC %>% filter(IBD == "UC"))[1]
summary(test_df_UC %>% filter(IBD == "Control"))
dim(test_df_UC %>% filter(IBD == "Control", C_KON == "K"))[1]/dim(test_df_UC %>% filter(IBD == "Control"))[1]

summary(test_df_CD %>% filter(IBD == "CD"))
dim(test_df_CD %>% filter(IBD == "CD", C_KON == "K"))[1]/dim(test_df_CD %>% filter(IBD == "CD"))[1]
summary(test_df_CD %>% filter(IBD == "Control"))
dim(test_df_CD %>% filter(IBD == "Control", C_KON == "K"))[1]/dim(test_df_CD %>% filter(IBD == "Control"))[1]

#Run model and extract results - CD
  
  #Turn the model
  test_df_CD$time2 <- -test_df_CD$sampling_to_diag
  f1 <- lmer(measure_transformed~  + C_KON + age + IBD * time2 + (1|lbnr)+(1|LABORATORIUM_IDCODE), data = test_df_CD)
  summary(f1)
  
    #QQ plot
  qqnorm(residuals(f1))
    
    #Residuals
  plot(f1)
  
    #Histogram
  ggplot()+
  geom_histogram(aes(x=residuals(f1)), bins=30)

  
#Run model and extract results - UC
  
  #Fit the model
  test_df_UC$time2 <- -test_df_UC$sampling_to_diag
  f1 <- lmer(measure_transformed~  + C_KON + age + IBD * time2 + (1|lbnr)+(1|LABORATORIUM_IDCODE), data = test_df_UC, control=lmerControl(optimizer = "bobyqa"))
  summary(f1)
  
    #QQ plot
  qqnorm(residuals(f1))
    
    #Residuals
  plot(f1)
  
    #Histogram
  ggplot()+
  geom_histogram(aes(x=residuals(f1)), bins=30)
```


Split into time intervals and test the association for each interval using LMM
LMM / time interval / feature:
```{r}
#CD:
# Add time interval column
max(test_df_CD$sampling_to_diag)/365.25
test_df_CD <- test_df_CD%>% mutate(time_interval = ifelse(sampling_to_diag<365.24, 1, ifelse(sampling_to_diag<2*365.25, 2, ifelse(sampling_to_diag<3*365.25,3, ifelse(sampling_to_diag<4*365.25,4, ifelse(sampling_to_diag<5*365.25,5, ifelse(sampling_to_diag<6*365.25,6,ifelse(sampling_to_diag<7*365.25,7, 8))))))))

# For each time interval!
res_CD <- data.frame()
for (i in 1:11){
    df_sub <- test_df_CD %>% filter(time_interval == i)
    form_CD <- "measure_transformed~ C_KON + age + IBD + (1|lbnr)+(1|LABORATORIUM_IDCODE)" 
    
    if(i %in% c(3, 5)){
      f_CD <- lmer(form_CD, data=df_sub, control=lmerControl(optimizer = "bobyqa"))
    }else{
      f_CD <- lmer(form_CD, data=df_sub)
    }
    
    
    print(summary(f_CD))
    ci <- confint(f_CD, parm = "IBDCD")
    print(ci)
    
    #Make result table
    tab <- summary(f_CD)
    res_sub <- data.frame(t(tab$coefficients["IBDCD",]))
    res_sub$time_interval <- i
    res_sub$ci1 <- ci[1]
    res_sub$ci2 <- ci[2]
    res_CD <- rbind(res_CD, res_sub)
    
}
colnames(res_CD)[1:5] <- c("Estimate", "SE", "df", "t_value", "p_value")
res_CD
write.table(res_CD, "F:/Projekter/FSEID00001565/Data/Subprojects/Marie_IBD_clinical_biomarker/results/VitD_restable_CD_onlyGP.txt", quote=F, row.names=F)

p <- ggplot(data=test_df_CD)+
    geom_boxplot(aes(x=as.factor(time_interval), y=measure_transformed, color=IBD))+
    ggtitle("Crohn's Disease: Vitamin D2 and D3")+theme_classic()+xlab("Years before diagnosis")+ylab("Measure (sqrt transformed)")


# UC:
# Add time interval column
max(test_df_UC$sampling_to_diag)/365.25
test_df_UC <- test_df_UC %>% mutate(time_interval = ifelse(sampling_to_diag<365.24, 1, ifelse(sampling_to_diag<2*365.25, 2, ifelse(sampling_to_diag<3*365.25,3, ifelse(sampling_to_diag<4*365.25,4, ifelse(sampling_to_diag<5*365.25,5, ifelse(sampling_to_diag<6*365.25,6,ifelse(sampling_to_diag<7*365.25,7,ifelse(sampling_to_diag<8*365.25,8, ifelse(sampling_to_diag<9*365.25,9,ifelse(sampling_to_diag<10*365.25, 10, 11)))))))))))

# For each time interval!
res_UC <- data.frame()
for (i in 1:11){
    df_sub <- test_df_UC %>% filter(time_interval == i)
    form_UC <- "measure_transformed~ C_KON + age + IBD + (1|lbnr)+(1|LABORATORIUM_IDCODE)" 
    
    f_UC <- lmer(form_UC, data=df_sub)
    
    print(summary(f_UC))
    ci <- confint(f_UC, parm = "IBDUC")
    print(ci)
    
    #Make result table
    tab <- summary(f_UC)
    res_sub <- data.frame(t(tab$coefficients["IBDUC",]))
    res_sub$time_interval <- i
    res_sub$ci1 <- ci[1]
    res_sub$ci2 <- ci[2]
    res_UC <- rbind(res_UC, res_sub)
}

colnames(res_UC)[1:5] <- c("Estimate", "SE", "df", "t_value", "p_value")
res_UC
write.table(res_UC, "F:/Projekter/FSEID00001565/Data/Subprojects/Marie_IBD_clinical_biomarker/results/VitD_restable_UC_onlyGP.txt", quote=F, row.names=F)

q <- ggplot(data=test_df_UC)+
    geom_boxplot(aes(x=as.factor(time_interval), y=measure_transformed, color=IBD))+
    ggtitle("Ulcerative colitis: Vitamin D2 and D3")+theme_classic()+xlab("Years before diagnosis")+ylab("Measure (sqrt transformed)")

ggarrange(p, q)

```

Sensitivity: Running the models only including one time point pr. case person:
```{r}
library(data.table)
#CD:

# For each time interval!
res_CD <- data.frame()
for (i in 1:11){
    #Subset dataframe:
    df_sub <- test_df_CD %>% filter(time_interval == i)
    df_case <- df_sub %>% filter(IBD=="CD")
    df_control <- df_sub %>% filter(IBD != "CD")
    df_case <- df_case[!duplicated(df_case$lbnr),]
    df_control <- data.table(df_control)
    df_control <- df_control[, head(.SD, 5), by="corresponding_case"]
    df_sub <- rbind(df_case, df_control)
    
    form_CD <- "measure_transformed~ C_KON + age + IBD+(1|LABORATORIUM_IDCODE)" 
    
    f_CD <- lmer(form_CD, data=df_sub)
    ci <- confint(f_CD, parm = "IBDCD")
    
    #Make result table
    tab <- summary(f_CD)
    res_sub <- data.frame(t(tab$coefficients["IBDCD",]))
    res_sub$time_interval <- i
    res_sub$ci1 <- ci[1]
    res_sub$ci2 <- ci[2]
    res_sub$n_cases <- dim(df_case)[1]
    res_sub$n_controls <- dim(df_control)[1]
    res_CD <- rbind(res_CD, res_sub)
}

colnames(res_CD)[1:5] <- c("Estimate", "SE", "df", "t_value", "p_value")
res_CD
write.table(res_CD, "F:/Projekter/FSEID00001565/Data/Subprojects/Marie_IBD_clinical_biomarker/results/One_VitD_restable_CD_onlyGP.txt", quote=F, row.names=F)

# UC:
# For each time interval!
res_UC <- data.frame()
for (i in 1:11){
    df_sub <- test_df_UC %>% filter(time_interval == i)
    df_case <- df_sub %>% filter(IBD=="UC")
    df_control <- df_sub %>% filter(IBD != "UC")
    df_case <- df_case[!duplicated(df_case$lbnr),]
    df_control <- data.table(df_control)
    df_control <- df_control[, head(.SD, 5), by="corresponding_case"]
    df_sub <- rbind(df_case, df_control)
    
    form_UC <- "measure_transformed~ C_KON + age + IBD+(1|LABORATORIUM_IDCODE)" 
    
    f_UC <- lmer(form_UC, data=df_sub)
    ci <- confint(f_UC, parm = "IBDUC")
    
    #Make result table
    tab <- summary(f_UC)
    res_sub <- data.frame(t(tab$coefficients["IBDUC",]))
    res_sub$time_interval <- i
    res_sub$ci1 <- ci[1]
    res_sub$ci2 <- ci[2]
    res_sub$n_cases <- dim(df_case)[1]
    res_sub$n_controls <- dim(df_control)[1]
    res_UC <- rbind(res_UC, res_sub)
}

colnames(res_UC)[1:5] <- c("Estimate", "SE", "df", "t_value", "p_value")
res_UC
write.table(res_UC, "F:/Projekter/FSEID00001565/Data/Subprojects/Marie_IBD_clinical_biomarker/results/One_VitD_restable_UC_onlyGP.txt", quote=F, row.names=F)
```


Fit CD vs. UC:
```{r}
res_total <- rbind(test_df_CD %>% filter(IBD == "CD"), test_df_UC %>% filter(IBD == "UC"))

f1 <- lmer(measure_transformed~  + C_KON + age + IBD * time2 + (1|lbnr)+(1|LABORATORIUM_IDCODE), data = res_total)
summary(f1)
  
#In time intervals:
res <- data.frame()
for (i in 1:11){
    df_sub <- res_total %>% filter(time_interval == i)
    form_UC <- "measure_transformed~ C_KON + age + IBD + (1|lbnr)+(1|LABORATORIUM_IDCODE)" 
    
    f_UC <- lmer(form_UC, data=df_sub)
    
    print(summary(f_UC))
    ci <- confint(f_UC, parm = "IBDUC")
    print(ci)
    
    #Make result table
    tab <- summary(f_UC)
    res_sub <- data.frame(t(tab$coefficients["IBDUC",]))
    res_sub$time_interval <- i
    res_sub$ci1 <- ci[1]
    res_sub$ci2 <- ci[2]
    res <- rbind(res, res_sub)
}

colnames(res)[1:5] <- c("Estimate", "SE", "df", "t_value", "p_value")
res
write.table(res, "F:/Projekter/FSEID00001565/Data/Subprojects/Marie_IBD_clinical_biomarker/results/VitD_UC_vs_CD_onlyGP.txt", quote=F, row.names=F)
```

## Sensitivity analysis: 2 IBD diagnoses
```{r}
test_df_CD <- read.csv("F:/Projekter/FSEID00001565/Data/Subprojects/Marie_IBD_clinical_biomarker/Sensitivity_Vitamin D1_CD_onlyGP.txt", sep=" ")
test_df_UC <- read.csv("F:/Projekter/FSEID00001565/Data/Subprojects/Marie_IBD_clinical_biomarker/Sensitivity_Vitamin D1_UC_onlyGP.txt", sep=" ")

test_df_CD <- test_df_CD %>% mutate(IBD_reg2 = fct_relevel(IBD_reg2, "Control", "CD"))

```

Check normality of the features and remove outliers:
```{r}
test_df_CD$measure_transformed <- as.numeric(test_df_CD$VALUE)
test_df_CD$VALUE[which(is.na(test_df_CD$measure_transformed))]

test_df_CD$VALUE <- gsub(",", ".", test_df_CD$VALUE)
test_df_UC$VALUE <- gsub(",", ".", test_df_UC$VALUE)

sub_CD <- test_df_CD %>% filter(as.numeric(VALUE)<10)
m_CD <- median(as.numeric(sub_CD$VALUE))
test_df_CD$VALUE <- gsub("<10", m_CD, test_df_CD$VALUE)
sub_CD <- test_df_CD %>% filter(as.numeric(VALUE)<12)
m_CD <- median(as.numeric(sub_CD$VALUE))
test_df_CD$VALUE <- gsub("<12", m_CD, test_df_CD$VALUE)
sub_CD <- test_df_CD %>% filter(as.numeric(VALUE)>375)
m_CD <- median(as.numeric(sub_CD$VALUE))
test_df_CD$VALUE <- gsub(">375", m_CD, test_df_CD$VALUE)
rm(sub_CD)

sub_UC <- test_df_UC %>% filter(as.numeric(VALUE)<10)
m_UC <- median(as.numeric(sub_UC$VALUE))
test_df_UC$VALUE <- gsub("<10", m_UC, test_df_UC$VALUE)
sub_UC <- test_df_UC %>% filter(as.numeric(VALUE)<12)
m_UC <- median(as.numeric(sub_UC$VALUE))
test_df_UC$VALUE <- gsub("<12", m_UC, test_df_UC$VALUE)
sub_UC <- test_df_UC %>% filter(as.numeric(VALUE)>375)
m_UC <- median(as.numeric(sub_UC$VALUE))
test_df_UC$VALUE <- gsub(">375", m_UC, test_df_UC$VALUE)
rm(sub_UC)

#Look at distribution:
  ggplot(test_df_CD)+
    geom_histogram(aes(x=as.numeric(VALUE)), bins=30)
  ggplot(test_df_CD)+
    geom_histogram(aes(x=log(as.numeric(VALUE)+1)), bins=30)
  ggplot(test_df_CD)+
    geom_histogram(aes(x=sqrt(as.numeric(VALUE))), bins=30)
  
  ggplot(test_df_UC)+
    geom_histogram(aes(x=as.numeric(VALUE)), bins=30)
  ggplot(test_df_UC)+
    geom_histogram(aes(x=sqrt(as.numeric(VALUE))), bins=30)
  
#Transform when necessary -> make a new column (measure_transformed)
test_df_CD$measure_transformed <- log(as.numeric(test_df_CD$VALUE)+1)
test_df_UC$measure_transformed <- log(as.numeric(test_df_UC$VALUE)+1)

test_df_CD <- test_df_CD %>% mutate(age =  as.numeric(as.Date(SAMPLINGDATE) - as.Date(D_FODDATO))/365.25)
test_df_UC <- test_df_UC %>% mutate(age =  as.numeric(as.Date(SAMPLINGDATE) - as.Date(D_FODDATO))/365.25)
```


Split into time intervals and test the association for each interval using LMM
LMM / time interval / feature:
```{r}
#CD:
# Add time interval column
max(test_df_CD$sampling_to_diag)/365.25
test_df_CD <- test_df_CD%>% mutate(time_interval = ifelse(sampling_to_diag<365.24, 1, ifelse(sampling_to_diag<2*365.25, 2, ifelse(sampling_to_diag<3*365.25,3, ifelse(sampling_to_diag<4*365.25,4, ifelse(sampling_to_diag<5*365.25,5, ifelse(sampling_to_diag<6*365.25,6,ifelse(sampling_to_diag<7*365.25,7,ifelse(sampling_to_diag<8*365.25,8, ifelse(sampling_to_diag<9*365.25,9,ifelse(sampling_to_diag<10*365.25, 10, 11)))))))))))

# For each time interval!
res_CD <- data.frame()
for (i in 1:11){
    df_sub <- test_df_CD %>% filter(time_interval == i)
    form_CD <- "measure_transformed~ C_KON + age + IBD_reg2 + (1|lbnr)+(1|LABORATORIUM_IDCODE)" 
    
    f_CD <- lmer(form_CD, data=df_sub)
    
    print(summary(f_CD))
    ci <- confint(f_CD, parm = "IBD_reg2CD")
    print(ci)
    
    #Make result table
    tab <- summary(f_CD)
    res_sub <- data.frame(t(tab$coefficients["IBD_reg2CD",]))
    res_sub$time_interval <- i
    res_sub$ci1 <- ci[1]
    res_sub$ci2 <- ci[2]
    res_CD <- rbind(res_CD, res_sub)
    
}
colnames(res_CD)[1:5] <- c("Estimate", "SE", "df", "t_value", "p_value")
res_CD
write.table(res_CD, "F:/Projekter/FSEID00001565/Data/Subprojects/Marie_IBD_clinical_biomarker/results/Sensitivity_VitD_restable_CD_onlyGP.txt", quote=F, row.names=F)


# UC:
# Add time interval column
max(test_df_UC$sampling_to_diag)/365.25
test_df_UC <- test_df_UC %>% mutate(time_interval = ifelse(sampling_to_diag<365.24, 1, ifelse(sampling_to_diag<2*365.25, 2, ifelse(sampling_to_diag<3*365.25,3, ifelse(sampling_to_diag<4*365.25,4, ifelse(sampling_to_diag<5*365.25,5, ifelse(sampling_to_diag<6*365.25,6,ifelse(sampling_to_diag<7*365.25,7,ifelse(sampling_to_diag<8*365.25,8, ifelse(sampling_to_diag<9*365.25,9,ifelse(sampling_to_diag<10*365.25, 10, 11)))))))))))

# For each time interval!
res_UC <- data.frame()
for (i in 1:11){
    df_sub <- test_df_UC %>% filter(time_interval == i)
    form_UC <- "measure_transformed~ C_KON + age + IBD_reg2 + (1|lbnr)+(1|LABORATORIUM_IDCODE)" 
    
    f_UC <- lmer(form_UC, data=df_sub)
    
    print(summary(f_UC))
    ci <- confint(f_UC, parm = "IBD_reg2UC")
    print(ci)
    
    #Make result table
    tab <- summary(f_UC)
    res_sub <- data.frame(t(tab$coefficients["IBD_reg2UC",]))
    res_sub$time_interval <- i
    res_sub$ci1 <- ci[1]
    res_sub$ci2 <- ci[2]
    res_UC <- rbind(res_UC, res_sub)
}

colnames(res_UC)[1:5] <- c("Estimate", "SE", "df", "t_value", "p_value")
res_UC
write.table(res_UC, "F:/Projekter/FSEID00001565/Data/Subprojects/Marie_IBD_clinical_biomarker/results/Sensitivity_VitD_restable_UC_onlyGP.txt", quote=F, row.names=F)

```

## Comorb sensitivity analysis:
```{r}
test_df_CD <- read.csv("F:/Projekter/FSEID00001565/Data/Subprojects/Marie_IBD_clinical_biomarker/Vitamin D1_CD_comorb_onlyGP.txt", sep=" ")
test_df_UC <- read.csv("F:/Projekter/FSEID00001565/Data/Subprojects/Marie_IBD_clinical_biomarker/Vitamin D1_UC_comorb_onlyGP.txt", sep=" ")

test_df_CD <- test_df_CD %>% mutate(IBD = fct_relevel(IBD, "Control", "CD"))
```

Check normality of the features and remove outliers:
```{r}
test_df_CD$VALUE <- gsub(",", ".", test_df_CD$VALUE)
test_df_UC$VALUE <- gsub(",", ".", test_df_UC$VALUE)

sub_CD <- test_df_CD %>% filter(as.numeric(VALUE)<10)
m_CD <- median(as.numeric(sub_CD$VALUE))
test_df_CD$VALUE <- gsub("<10", m_CD, test_df_CD$VALUE)
sub_CD <- test_df_CD %>% filter(as.numeric(VALUE)<12)
m_CD <- median(as.numeric(sub_CD$VALUE))
test_df_CD$VALUE <- gsub("<12", m_CD, test_df_CD$VALUE)
sub_CD <- test_df_CD %>% filter(as.numeric(VALUE)>375)
m_CD <- median(as.numeric(sub_CD$VALUE))
test_df_CD$VALUE <- gsub(">375", m_CD, test_df_CD$VALUE)
rm(sub_CD)

sub_UC <- test_df_UC %>% filter(as.numeric(VALUE)<10)
m_UC <- median(as.numeric(sub_UC$VALUE))
test_df_UC$VALUE <- gsub("<10", m_UC, test_df_UC$VALUE)
sub_UC <- test_df_UC %>% filter(as.numeric(VALUE)<12)
m_UC <- median(as.numeric(sub_UC$VALUE))
test_df_UC$VALUE <- gsub("<12", m_UC, test_df_UC$VALUE)
sub_UC <- test_df_UC %>% filter(as.numeric(VALUE)>375)
m_UC <- median(as.numeric(sub_UC$VALUE))
test_df_UC$VALUE <- gsub(">375", m_CD, test_df_UC$VALUE)
rm(sub_UC)

#Transform when necessary -> make a new column (measure_transformed)
test_df_CD$measure_transformed <- log(as.numeric(test_df_CD$VALUE)+1)
test_df_UC$measure_transformed <- log(as.numeric(test_df_UC$VALUE)+1)
  
# Investigate covariates in the two cohorts (for each NPU)
CD_cases <- test_df_CD %>% filter(IBD == "CD")
CD_controls <- test_df_CD %>% filter(IBD != "CD")
UC_cases <- test_df_UC %>% filter(IBD == "UC")
UC_controls <- test_df_UC %>% filter(IBD != "UC")

summary(CD_cases)
summary(CD_controls)
summary(UC_cases)
summary(UC_controls)

rm(CD_cases, CD_controls, UC_cases, UC_controls)

test_df_CD <- test_df_CD %>% mutate(age =  as.numeric(as.Date(SAMPLINGDATE) - as.Date(D_FODDATO))/365.25)
test_df_UC <- test_df_UC %>% mutate(age =  as.numeric(as.Date(SAMPLINGDATE) - as.Date(D_FODDATO))/365.25)
```


Split into time intervals and test the association for each interval using LMM
LMM / time interval / feature:
```{r}
#CD:
# Add time interval column
max(test_df_CD$sampling_to_diag)/365.25
test_df_CD <- test_df_CD%>% mutate(time_interval = ifelse(sampling_to_diag<365.24, 1, ifelse(sampling_to_diag<2*365.25, 2, ifelse(sampling_to_diag<3*365.25,3, ifelse(sampling_to_diag<4*365.25,4, ifelse(sampling_to_diag<5*365.25,5, ifelse(sampling_to_diag<6*365.25,6,ifelse(sampling_to_diag<7*365.25,7, 8))))))))

# For each time interval!
res_CD <- data.frame()
for (i in 1:8){
    df_sub <- test_df_CD %>% filter(time_interval == i)
    form_CD <- "measure_transformed~ C_KON + age + comorb_score + IBD + (1|lbnr)+(1|LABORATORIUM_IDCODE)" 
    
    f_CD <- lmer(form_CD, data=df_sub)
    
    print(summary(f_CD))
    ci <- confint(f_CD, parm = "IBDCD")
    print(ci)
    
    #Make result table
    tab <- summary(f_CD)
    res_sub <- data.frame(t(tab$coefficients["IBDCD",]))
    res_sub$time_interval <- i
    res_sub$ci1 <- ci[1]
    res_sub$ci2 <- ci[2]
    res_CD <- rbind(res_CD, res_sub)
    
}
colnames(res_CD)[1:5] <- c("Estimate", "SE", "df", "t_value", "p_value")
res_CD
write.table(res_CD, "F:/Projekter/FSEID00001565/Data/Subprojects/Marie_IBD_clinical_biomarker/results/VitD_restable_CD_comorb_onlyGP.txt", quote=F, row.names=F)


# UC:
# Add time interval column
max(test_df_UC$sampling_to_diag)/365.25
test_df_UC <- test_df_UC %>% mutate(time_interval = ifelse(sampling_to_diag<365.24, 1, ifelse(sampling_to_diag<2*365.25, 2, ifelse(sampling_to_diag<3*365.25,3, ifelse(sampling_to_diag<4*365.25,4, ifelse(sampling_to_diag<5*365.25,5, ifelse(sampling_to_diag<6*365.25,6,ifelse(sampling_to_diag<7*365.25,7,ifelse(sampling_to_diag<8*365.25,8, ifelse(sampling_to_diag<9*365.25,9,ifelse(sampling_to_diag<10*365.25, 10, 11)))))))))))

# For each time interval!
res_UC <- data.frame()
for (i in 1:11){
    df_sub <- test_df_UC %>% filter(time_interval == i)
    form_UC <- "measure_transformed~ C_KON + age + comorb_score + IBD + (1|lbnr)+(1|LABORATORIUM_IDCODE)" 
    
    if(i == 4){
      f_UC <- lmer(form_UC, data=df_sub, control = lmerControl(optimizer = "bobyqa"))
    } else{
      f_UC <- lmer(form_UC, data=df_sub)
    }
    
    
    print(summary(f_UC))
    ci <- confint(f_UC, parm = "IBDUC")
    print(ci)
    
    #Make result table
    tab <- summary(f_UC)
    res_sub <- data.frame(t(tab$coefficients["IBDUC",]))
    res_sub$time_interval <- i
    res_sub$ci1 <- ci[1]
    res_sub$ci2 <- ci[2]
    res_UC <- rbind(res_UC, res_sub)
}

colnames(res_UC)[1:5] <- c("Estimate", "SE", "df", "t_value", "p_value")
res_UC
write.table(res_UC, "F:/Projekter/FSEID00001565/Data/Subprojects/Marie_IBD_clinical_biomarker/results/VitD_restable_UC_comorb_onlyGP.txt", quote=F, row.names=F)

```
